Now we can build discrete tikzpictures.
If the user asks to build a discrete tikzpicture, lunatikz will build
that file usually then, lunatikz will take this individual pdf files,
and build the user asked pdf with the needed margin.

```
\documentclass [path/to/main/file] {subfiles}

\begin{document}

    \standaloneconfig{margin = {user defined margin}}  <----+
                                                            |
    \begin{tikzpicture}                                     |
    \node [inner sep = 0pt]                                 |
    {\includegraphics {path/to/previously/build/pdf}};      |
    \end{tikzpicture}                                  <----+
                                                            |
    ...                            For each file -----------+
    ...
    many more just like above
    ...

\end{document}

```

lunatikz will create a standalone_sub file as above, adding all the
pdf files. And compiles this using `pdflatex` and splits it using
`pdftk`, and places them in the current directory with the style
appended to it.