bug fix: pwd and readlink would return two differet paths (On Android)
This issue is not fixed entirely, and not exclusivly affects Termux,
this could affect any lunatikz project that has a soft link to another
directory. In that case the `readlink` command will return the actual
directory, instead of returning the linked directory.

Currently this is fixed if the project directory is entirely in the
internal storage directory or is entirely in the termux private directory
where `readlink` will returns the proper directory.