bug fix: fixed stripping the comments.
First of all lunatikz will strip every \\%<some comment> into \\ then it
will convert every comment that is not in the form \%<not comment>. ie,
it will look for a "<a single character that is not '/'><'%'><some comment>",
then it will replace it will the whole thing with "<a single character that
is not '/'>", this will ensure that we don't accidently remove something
like "\%" which is not a start of a comment. But this will slip every
comment that is achored at the start of the line.

Now we will remove those leftover comment.