added subtikzpicture.sty
Which is circuitikz's subcircuits, but on steroids. `subtikzpicture` implements
a `subtikzpicturedef` which is similar to `ctikzsubcircuitdef` but allows
the definition of `subtikzpicture` just like a normal `tikzpicture`. And
it also allows the proper working of `subfix` inside the `subtikzpicture`.
As long as the `subfix` file exists relative to the `tikzpic`, lunatikz
can build it from anywhere inside the lunatikz project directory tree.