bug fix: special chars weren't properly escaped in :match and :gsub
If chars like `-`, and `.` weren't properly escaped, it won't properly
:match or :gsub. Currently, escape_special_chars() escapes these
characters. But this seems like a bad design. It is possible that we
have missed escaping of these characters somewhere else, and possible
to forget to escape these characters in the future.

It seems like the time has come to rewrite lunatikz in a reliable language.